rewrite the storageRegion objects

+++ FS_Server/src/main/java/com/bonree/brfs/duplication/storageregion/StorageRegionIdBuilder.java
@@ -1,5 +1,5 @@
 package com.bonree.brfs.duplication.storageregion;
 
 public interface StorageRegionIdBuilder {
-	short createRegionId() throws Exception;
+    int createRegionId() throws Exception;
 }
